refactor: Moved utilities to separate folders. Moved styles to a separate folder.
Moved utilities to separate folders. Moved styles to a separate folder.
