+++ my-app/src/const.js
@@ -1,0 +1,6 @@
+export const UP_ARROW = 38;
+export const DOWN_ARROW = 40;
+export const LEFT_ARROW = 37;
+export const RIGHT_ARROW = 39;
+
+
