+++ my-app/src/style/style.js
@@ -1,0 +1,50 @@
+const style = {
+    blockStyle: {
+        height: 80,
+        width: 80,
+        background: "lightgray",
+        margin: 3,
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+        fontSize: 45,
+        fontWeight: "800",
+        color: "white",
+    },
+    newGameButton: {
+        padding: 10,
+        background: "#846F5B",
+        color: "#F8F5F0",
+        width: 95,
+        borderRadius: 7,
+        fontWeight: "900",
+        marginLeft: "auto",
+        marginBottom: "auto",
+        cursor: "pointer",
+    },
+    tryAgainButton: {
+        padding: 10,
+        background: "#846F5B",
+        color: "#F8F5F0",
+        width: 80,
+        borderRadius: 7,
+        fontWeight: "900",
+        cursor: "pointer",
+        margin: "auto",
+        marginTop: 20,
+    },
+    gameOverOverlay: {
+        position: "absolute",
+        height: "100%",
+        width: "100%",
+        left: 0,
+        top: 0,
+        borderRadius: 5,
+        background: "rgba(238,228,218,.5)",
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+    },
+};
+
+export default style;
