+++ my-app/src/utils/util.js
@@ -1,0 +1,14 @@
+import { useEffect } from "react";
+
+const useEvent = (event, handler, passive = false) => {
+    useEffect(() => {
+
+        window.addEventListener(event, handler, passive);
+
+        return function cleanup() {
+            window.removeEventListener(event, handler);
+        };
+    });
+};
+export default useEvent;
+
